+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Changelog
 
+## Unreleased
+
 ## 0.3.14 - 2026-06-11
 
 ### Changed